Udpated GetMZmineFeature to more flexibly match data columns between metadata and feature abunance matrix.
This versions keeps check.names = F as the current default. I think this is important because we will never know what people will name there samples so it's better to just match them directly from the metadata. It also removes the data: and the :area from the column names which will help with downstream graphs etc.

I also added checks in case the matching does not work correctly and the feature abundance table has a different number of samples than the metadata